Making the SVG resize according to the viewport dimensions
It's Scalable vector graphics, so, let's make it scalable!

With this commit, the main simulator SVG will fit the device
width/height, respecting the original aspect ratio, and keeping it
centered. It's now easy to enjoy the simulator in a large screen, or
even on smaller windows without having to scroll.